+ changed type of constants in ilp terms to Double + added ILPLinearFunction and ILPNestedLinearFunction to model objective functions

+++ org.emoflon.roam.core/src/org/emoflon/roam/core/RoamMappingConstraint.java
@@ -25,8 +25,8 @@ public abstract class RoamMappingConstraint<CONTEXT extends RoamMapping> extends
 	
 	@Override
 	public ILPConstraint<Integer> buildConstraint(final CONTEXT context) {
-		Integer constTerm = buildConstantTerm(context);
-		List<ILPTerm<Integer, Integer>> terms = buildVariableTerms(context);
+		double constTerm = buildConstantTerm(context);
+		List<ILPTerm<Integer, Double>> terms = buildVariableTerms(context);
 		return new ILPConstraint<Integer>(constTerm, constraint.getExpression().getOperator(), terms);
 	}
